\name{VARsignR}

\alias{VARsignR}
\alias{VARsignR-package}

\docType{package}

\title{Estimating VARs using sign restrictions}

\description{
VARsignR is a collection of R routines for estimating and identifying structural shocks using sign restrictions in Bayesian Vector Autoregression (BVAR) models.
}

\details{
\tabular{ll}{
Package: \tab VARsignR\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2015-12-19\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab Yes\cr
Depends: \tab HI, mvnfast, minqa\cr }
}

\section{How to cite this package}{

 \preformatted{@Manual{Danne2015,
 author       = {Christian Danne},
 title        = {{VARsignR}: Estimating VARs using sign restrictions.},
 year         = {2015},
 note         = {R package version 0.1.2.}}}
\emph{}
}

\section{License}{
VARsignR is licensed under the GNU General Public License (GPL) version >=3. You can redistribute and/or modify it under the terms of the GNU GPL as published by the Free Software Foundation. The package comes without any warranty, without even the implied warranty of merchantability, or fitness for a particular purpose. See the GNU GPL for more details.
}

\author{Christian Danne}
